+++ delfin/tests/unit/task_manager/scheduler/schedulers/telemetry/test_performance_collection_handler.py
@@ -28,6 +28,7 @@ from delfin.task_manager.scheduler.schedulers.telemetry. \
     PerformanceCollectionHandler
 
 fake_task_id = 43
+fake_executor = 'node1'
 fake_storage_id = '12c2d52f-01bc-41f5-b73f-7abf6f38a2a6'
 fake_telemetry_job = {
     Task.id.name: 2,
@@ -35,7 +36,8 @@ fake_telemetry_job = {
     Task.args.name: {},
     Task.interval.name: 10,
     Task.deleted.name: False,
-    Task.method.name: constants.TelemetryCollection.PERFORMANCE_TASK_METHOD
+    Task.method.name: constants.TelemetryCollection.PERFORMANCE_TASK_METHOD,
+    Task.executor.name: fake_executor
 }
 
 fake_deleted_telemetry_job = {
@@ -44,7 +46,8 @@ fake_deleted_telemetry_job = {
     Task.args.name: {},
     Task.interval.name: 10,
     Task.deleted.name: True,
-    Task.method.name: constants.TelemetryCollection.PERFORMANCE_TASK_METHOD
+    Task.method.name: constants.TelemetryCollection.PERFORMANCE_TASK_METHOD,
+    Task.executor.name: fake_executor
 }
 
 
@@ -57,7 +60,8 @@ class TestPerformanceCollectionHandler(test.TestCase):
     @mock.patch.object(db, 'task_get',
                        mock.Mock(return_value=fake_telemetry_job))
     @mock.patch('delfin.db.task_update')
-    @mock.patch('delfin.task_manager.rpcapi.TaskAPI.collect_telemetry')
+    @mock.patch('delfin.task_manager.tasks.telemetry'
+                '.PerformanceCollectionTask.collect')
     def test_performance_collection_success(self, mock_collect_telemetry,
                                             mock_task_update):
         mock_collect_telemetry.return_value = TelemetryTaskStatus. \
@@ -74,7 +78,8 @@ class TestPerformanceCollectionHandler(test.TestCase):
     @mock.patch.object(db, 'task_get',
                        mock.Mock(return_value=fake_telemetry_job))
     @mock.patch('delfin.db.failed_task_create')
-    @mock.patch('delfin.task_manager.rpcapi.TaskAPI.collect_telemetry')
+    @mock.patch('delfin.task_manager.tasks.telemetry'
+                '.PerformanceCollectionTask.collect')
     def test_performance_collection_failure(self, mock_collect_telemetry,
                                             mock_failed_task_create):
         mock_collect_telemetry.return_value = TelemetryTaskStatus. \
@@ -91,7 +96,8 @@ class TestPerformanceCollectionHandler(test.TestCase):
     @mock.patch.object(db, 'task_get',
                        mock.Mock(return_value=fake_deleted_telemetry_job))
     @mock.patch('delfin.db.task_update')
-    @mock.patch('delfin.task_manager.rpcapi.TaskAPI.collect_telemetry')
+    @mock.patch('delfin.task_manager.tasks.telemetry'
+                '.PerformanceCollectionTask.collect')
     def test_performance_collection_deleted_storage(self,
                                                     mock_collect_telemetry,
                                                     mock_task_update):
@@ -108,14 +114,16 @@ class TestPerformanceCollectionHandler(test.TestCase):
         self.assertEqual(mock_task_update.call_count, 0)
 
     @mock.patch('delfin.db.task_get', task_not_found_exception)
-    @mock.patch('delfin.task_manager.rpcapi.TaskAPI.collect_telemetry')
+    @mock.patch('delfin.task_manager.tasks.telemetry'
+                '.PerformanceCollectionTask.collect')
     def test_deleted_storage_exception(self,
                                        mock_collect_telemetry):
         ctx = context.get_admin_context()
         perf_collection_handler = PerformanceCollectionHandler(ctx,
                                                                fake_task_id,
                                                                fake_storage_id,
-                                                               "", 100)
+                                                               "", 100,
+                                                               fake_executor)
         perf_collection_handler()
 
         # Verify that collect telemetry for deleted storage
